+++ mirrord-config/src/lib.rs
@@ -9,6 +9,7 @@ pub mod fs;
 pub mod incoming;
 pub mod network;
 pub mod outgoing;
+pub mod target;
 pub mod util;
 
 use std::path::Path;
@@ -18,7 +19,7 @@ use serde::Deserialize;
 
 use crate::{
     agent::AgentFileConfig, config::source::MirrordConfigSource, feature::FeatureFileConfig,
-    util::VecOrSingle,
+    target::TargetFileConfig, util::VecOrSingle,
 };
 
 /// This is the root struct for mirrord-layer's configuration
@@ -33,11 +34,9 @@ pub struct LayerFileConfig {
     #[config(env = "MIRRORD_SKIP_PROCESSES")]
     pub skip_processes: Option<VecOrSingle<String>>,
 
-    #[config(env = "MIRRORD_IMPERSONATED_TARGET")]
-    pub target: Option<String>,
-
-    #[config(env = "MIRRORD_TARGET_NAMESPACE")]
-    pub target_namespace: Option<String>,
+    #[serde(default)]
+    #[config(nested)]
+    pub target: TargetFileConfig,
 
     #[cfg_attr(feature = "schema", schemars(skip))]
     /// IP:PORT to connect to instead of using k8s api, for testing purposes.
@@ -87,6 +86,7 @@ mod tests {
         incoming::IncomingConfig,
         network::NetworkFileConfig,
         outgoing::OutgoingFileConfig,
+        target::{PodTarget, Target, TargetFileConfig},
         util::ToggleableConfig,
     };
 
@@ -112,8 +112,10 @@ mod tests {
                     r#"
                     {
                         "accept_invalid_certificates": false,
-                        "target": "pod/test-service-abcdefg-abcd",
-                        "target_namespace": "default",
+                        "target": {
+                            "path": "pod/test-service-abcdefg-abcd",
+                            "namespace": "default"
+                        },
                         "agent": {
                             "log_level": "info",
                             "namespace": "default",
@@ -140,8 +142,10 @@ mod tests {
                 ConfigType::Toml => {
                     r#"
                     accept_invalid_certificates = false
-                    target = "pod/test-service-abcdefg-abcd"
-                    target_namespace = "default"
+
+                    [target]
+                    path = "pod/test-service-abcdefg-abcd"
+                    namespace = "default"
 
                     [agent]
                     log_level = "info"
@@ -161,14 +165,15 @@ mod tests {
 
                     [feature.network.outgoing]
                     tcp = true
-                    udp = false                    
+                    udp = false
                     "#
                 }
                 ConfigType::Yaml => {
                     r#"
                     accept_invalid_certificates: false
-                    target: "pod/test-service-abcdefg-abcd"
-                    target_namespace: "default"
+                    target: 
+                        path: "pod/test-service-abcdefg-abcd"
+                        namespace: "default"
 
                     agent:
                         log_level: "info"
@@ -228,8 +233,13 @@ mod tests {
             accept_invalid_certificates: Some(false),
             connect_agent_name: None,
             connect_agent_port: None,
-            target: Some("pod/test-service-abcdefg-abcd".to_owned()),
-            target_namespace: Some("default".to_owned()),
+            target: TargetFileConfig::Advanced {
+                path: Some(Target::Pod(PodTarget {
+                    pod: "test-service-abcdefg-abcd".to_owned(),
+                    container: None,
+                })),
+                namespace: Some("default".to_owned()),
+            },
             skip_processes: None,
             agent: AgentFileConfig {
                 log_level: Some("info".to_owned()),